# docs/openapi.yaml
openapi: 3.0.0
info:
  title: NyaySetu Unified API
  version: 1.0.0
  description: REST API specification for the NyaySetu Digital Judiciary Platform. Covers Spring Boot backend services and Python microservices.
servers:
  - url: http://localhost:8080
    description: Spring Boot Backend Server
  - url: http://localhost:8001
    description: Python Services (NLP Orchestrator / LawGPT)
paths:
...
  /api/v1/documents:
    get:
      tags:
        - Document Management
      summary: "Execute getUserDocuments"
      description: " (Served by Spring Boot backend)"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
  /api/v1/documents/case/{caseId}:
    get:
      tags:
        - Document Management
      summary: "Execute getCaseDocuments"
      description: " (Served by Spring Boot backend)"
      parameters:
        - name: caseId
          in: path
          required: true
          schema:
            type: string
          description: "The caseId identifier"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
  /api/v1/documents/upload:
    post:
      tags:
        - Document Management
      summary: "Execute uploadDocument"
      description: " (Served by Spring Boot backend)"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
  /api/v1/documents/user/cases:
    get:
      tags:
        - Document Management
      summary: "Execute getUserCases"
      description: " (Served by Spring Boot backend)"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
  /api/v1/documents/{id}:
    delete:
      tags:
        - Document Management
      summary: "Execute deleteDocument"
      description: " (Served by Spring Boot backend)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "The id identifier"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
    get:
      tags:
        - Document Management
      summary: "Execute getDocument"
      description: " (Served by Spring Boot backend)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "The id identifier"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
  /api/v1/documents/{id}/analysis:
    get:
      tags:
        - Document Management
      summary: "Get AI analysis for a document"
      description: "Get AI analysis for a document (Served by Spring Boot backend)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "The id identifier"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
  /api/v1/documents/{id}/analyze:
    post:
      tags:
        - Document Management
      summary: "Trigger AI analysis for a document"
      description: "Trigger AI analysis for a document (Served by Spring Boot backend)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "The id identifier"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
  /api/v1/documents/{id}/certificate:
    get:
      tags:
        - Document Management
      summary: "Download Section 63(4) Evidence Certificate for a document"
      description: "Download Section 63(4) Evidence Certificate for a document (Served by Spring Boot backend)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "The id identifier"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
  /api/v1/documents/{id}/download:
    get:
      tags:
        - Document Management
      summary: "Execute downloadDocument"
      description: " (Served by Spring Boot backend)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "The id identifier"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
  /api/v1/documents/{id}/has-analysis:
    get:
      tags:
        - Document Management
      summary: "Check if document has analysis"
      description: "Check if document has analysis (Served by Spring Boot backend)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "The id identifier"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
  /api/v1/documents/{id}/verify-hash:
    get:
      tags:
        - Document Management
      summary: "Verify document hash (SHA-256) againts stored fingerprint"
      description: "Verify document hash (SHA-256) againts stored fingerprint (Served by Spring Boot backend)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: "The id identifier"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "Successful operation"
        '401':
          description: "Unauthorized"
        '403':
          description: "Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT Token from /api/v1/auth/login"